streamguard-backend/requirements.txt
fastapi
uvicorn[standard]
websockets
openai>=1.0.0
python-dotenv
python-multipart
requests
protobuf
